+++ crates/consensus/src/lib.rs
@@ -19,6 +19,7 @@ pub use transaction::{
     DEPOSIT_TX_TYPE_ID,
 };
 
+pub mod eip1559;
 pub mod hardforks;
 pub use hardforks::Hardforks;
