+++ crates/rpc-types-engine/src/lib.rs
@@ -25,6 +25,3 @@ mod superchain;
 pub use superchain::{
     ProtocolVersion, ProtocolVersionError, ProtocolVersionFormatV0, SuperchainSignal,
 };
-
-mod error;
-pub use error::EIP1559ParamError;
